icons

+++ ZKL/ViewControllers/HomeControlVC.m
@@ -125,7 +125,7 @@
 {
     CGRect rect = CGRectMake(0, 0, kScreenWidth, kScreenHeight);
     UIImageView *imageView = [[UIImageView alloc] initWithFrame:rect];
-    [imageView setImage:[UIImage     imageNamed:[NSString stringWithFormat:@"引导页%d",index+1]]];
+    [imageView setImage:[UIImage     imageNamed:[NSString stringWithFormat:@"引导页%ld",(long)index+1]]];
     imageView.contentMode = UIViewContentModeScaleAspectFit;
     return imageView;
 }
